move logger utilities to lite
