+++ tests/tests_pytorch/loggers/test_logger.py
@@ -21,12 +21,13 @@ import numpy as np
 import pytest
 import torch
 
+from lightning_fabric.utilities.logger import _convert_params, _sanitize_params
 from pytorch_lightning import Trainer
 from pytorch_lightning.demos.boring_classes import BoringDataModule, BoringModel
 from pytorch_lightning.loggers import Logger, TensorBoardLogger
 from pytorch_lightning.loggers.logger import DummyExperiment, DummyLogger
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
-from pytorch_lightning.utilities.logger import _convert_params, _sanitize_params, _scan_checkpoints
+from pytorch_lightning.utilities.logger import _scan_checkpoints
 from pytorch_lightning.utilities.rank_zero import rank_zero_only
 from tests_pytorch.helpers.runif import RunIf
